+++ pallets/creditcoin/src/ocw/tasks/verify_transfer.rs
@@ -212,77 +212,3 @@ impl<T: Config> crate::Pallet<T> {
 		Ok(timestamp)
 	}
 }
-
-type UnverifiedTransferT<T> = UnverifiedTransfer<
-	<T as SystemConfig>::AccountId,
-	<T as SystemConfig>::BlockNumber,
-	<T as SystemConfig>::Hash,
-	<T as TimestampConfig>::Moment,
->;
-
-impl<T: Config> OffchainVerification<T> for UnverifiedTransferT<T> {
-	type Output = Option<T::Moment>;
-
-	fn verify(&self) -> VerificationResult<Self::Output> {
-		crate::Pallet::<T>::verify_transfer_ocw(self)
-	}
-}
-
-use crate::ocw::errors::SchedulerError;
-use crate::ocw::tasks::OffchainVerification;
-use crate::types::concatenate;
-use crate::TaskOutput;
-use crate::TransferId;
-use frame_system::Config as SystemConfig;
-use pallet_offchain_task_scheduler::tasks::error::TaskError;
-use pallet_offchain_task_scheduler::tasks::TaskV2;
-use pallet_timestamp::Config as TimestampConfig;
-use sp_runtime::traits::Hash;
-
-impl<T: Config> TaskV2<T> for UnverifiedTransferT<T>
-where
-	UnverifiedTransferT<T>: OffchainVerification<T>,
-	<UnverifiedTransferT<T> as OffchainVerification<T>>::Output: Into<Option<T::Moment>>,
-{
-	type Call = crate::pallet::Call<T>;
-	type EvaluationError = VerificationFailureCause;
-	type SchedulerError = SchedulerError;
-
-	fn to_id(&self) -> T::Hash {
-		let Transfer { blockchain, tx_id, .. } = &self.transfer;
-		let key = concatenate!(blockchain.as_bytes().as_ref(), tx_id.as_slice());
-		T::Hashing::hash(&key)
-	}
-
-	fn persistence_call(
-		&self,
-		deadline: T::BlockNumber,
-		id: &T::Hash,
-	) -> Result<Self::Call, TaskError<Self::EvaluationError, Self::SchedulerError>> {
-		use crate::ocw::OffchainError::*;
-		match self.verify() {
-			Ok(timestamp) => {
-				let transfer = self.clone().into_output::<T>(timestamp.into());
-				let id = TransferId::from(*id);
-				Ok(Self::Call::persist_task_output {
-					deadline,
-					task_output: TaskOutput::from((id, transfer)),
-				})
-			},
-			Err(InvalidTask(cause)) if cause.is_fatal() => {
-				log::warn!("Failed to verify pending task {:?} : {:?}", self, cause);
-				let id = TransferId::from(*id);
-				Ok(Self::Call::fail_task { deadline, task_id: id.into(), cause })
-			},
-			Err(InvalidTask(e)) => Err(TaskError::Evaluation(e)),
-			Err(NoRpcUrl(e)) => Err(TaskError::Scheduler(e.into())),
-			Err(RpcError(e)) => Err(TaskError::Scheduler(e.into())),
-			Err(IncorrectChainId) => Err(TaskError::Scheduler(SchedulerError::IncorrectChainId)),
-		}
-	}
-
-	fn is_persisted(id: &T::Hash) -> bool {
-		let id = TransferId::from(*id);
-		crate::pallet::Transfers::<T>::contains_key(&id)
-	}
-}
